Fix critical RISC-V bugs and eliminate all compiler warnings
Critical bug fixes:
- switch.S: Fix stack offset (24, not 0) for struct thread.stack field
- stdint.h: Fix uintptr_t to be 64-bit on RISC-V (was causing address truncation)
- intr.c: Fix intr_context() to track actual trap handler execution
- palloc.c: Use identity-mapped addresses consistently for RISC-V
- mmu.c: Add MMIO region mapping (L2[0]) for PLIC/UART access
- trap.S: Initialize sscratch to 0 for S-mode trap handling
- switch.S: Call thread_switch_tail instead of undefined schedule_tail

Warning fixes (14 -> 0):
- Add missing prototypes to intr.h, virtio.h
- Create devices.h with serial/shutdown function declarations
- Remove unused early_panic/early_memset from init.c
- Fix type comparison in stdio.c (signedness, format specifier)
- Fix pointer/integer comparison in debug.c

New files:
- arch/riscv64/boot.h: Boot variable declarations
- arch/riscv64/devices.c: SBI-based device stubs
- arch/riscv64/devices.h: Device function prototypes
- riscv/Make.vars: RISC-V build configuration

The kernel now boots successfully on QEMU with working threading,
context switching, and timer interrupts at 100 Hz.

+++ src/lib/kernel/debug.c
@@ -94,7 +94,7 @@ static void print_stacktrace(struct thread* t, void* aux UNUSED) {
     frame = (void**)saved_frame->ebp;
     retaddr = (void*)saved_frame->eip;
 #elif defined(ARCH_RISCV64)
-    if (t->stack == (uint8_t*)t + PGSIZE || saved_frame->ra == switch_entry) {
+    if (t->stack == (uint8_t*)t + PGSIZE || saved_frame->ra == (uintptr_t)switch_entry) {
       printf(" thread was never scheduled.\n");
       return;
     }
